Change the "Based on work" link
The source as is displays the single link on the browser (tested on Firefox and Chrome)  "[https://github.com/HaxeFoundation/HaxeManual](https://github.com/HaxeFoundation/HaxeManual)". So, as expected, this redirects the user to a 404 page on github. I don't think this is intendend.

I changed the link description to "Haxe Manual on GitHub", but it's just a suggestion. You may have a better description for it.